Fixed a fault that prevented decimal values for CSS margins being converted correctly.

+++ src/heat.js
@@ -780,7 +780,7 @@
         }   
         
         if ( toNumber ) {
-            value = parseInt( value, 10 );
+            value = parseFloat( value, 10 );
         }
 
         return value;
